test: restore shared site-binding helper
